--- a/encoding/fuzz_test.go
+++ b/encoding/fuzz_test.go
@@ -5,10 +5,8 @@ package encoding_test
 
 import (
 	"bytes"
-	"encoding/binary"
 	"errors"
 	"io"
-	"math/rand"
 	"testing"
 
 	"github.com/segmentio/parquet-go/deprecated"
@@ -19,6 +17,7 @@ import (
 	"github.com/segmentio/parquet-go/encoding/rle"
 	"github.com/segmentio/parquet-go/format"
 	"github.com/segmentio/parquet-go/internal/bits"
+	"github.com/segmentio/parquet-go/internal/fuzzing"
 )
 
 func fuzzEncoding(fuzz func(e encoding.Encoding)) {
@@ -73,14 +72,14 @@ func fuzzEncoding(fuzz func(e encoding.Encoding)) {
 func FuzzAllEncoding(f *testing.F) {
 	f.Fuzz(func(t *testing.T, input []byte, size int) {
 		fuzzEncoding(func(e encoding.Encoding) {
-			fuzzBooleanDecoding(t, makeRandBoolean(input, size), e)
+			fuzzBooleanDecoding(t, fuzzing.MakeRandBoolean(input, size), e)
 			fuzzByteArrayDecoding(t, input, e)
 			fuzzFixedLenByteArrayDecoding(t, size, input, e)
-			fuzzFloatDecoding(t, makeRandFloat(input, size), e)
-			fuzzDoubleDecoding(t, makeRandDouble(input, size), e)
-			fuzzInt32Decoding(t, makeRandInt32(input, size), e)
-			fuzzInt64Decoding(t, makeRandInt64(input, size), e)
-			fuzzInt96Decoding(t, makeRandInt96(input, size), e)
+			fuzzFloatDecoding(t, fuzzing.MakeRandFloat(input, size), e)
+			fuzzDoubleDecoding(t, fuzzing.MakeRandDouble(input, size), e)
+			fuzzInt32Decoding(t, fuzzing.MakeRandInt32(input, size), e)
+			fuzzInt64Decoding(t, fuzzing.MakeRandInt64(input, size), e)
+			fuzzInt96Decoding(t, fuzzing.MakeRandInt96(input, size), e)
 		})
 	})
 }
@@ -88,7 +87,7 @@ func FuzzAllEncoding(f *testing.F) {
 func FuzzBooleanEncoding(f *testing.F) {
 	f.Fuzz(func(t *testing.T, input []byte, count int) {
 		fuzzEncoding(func(e encoding.Encoding) {
-			fuzzBooleanDecoding(t, makeRandBoolean(input, count), e)
+			fuzzBooleanDecoding(t, fuzzing.MakeRandBoolean(input, count), e)
 		})
 	})
 }
@@ -112,8 +111,8 @@ func FuzzFixedLenByteArrayEncoding(f *testing.F) {
 func FuzzFloatEncoding(f *testing.F) {
 	f.Fuzz(func(t *testing.T, size int, input []byte) {
 		fuzzEncoding(func(e encoding.Encoding) {
-			fuzzFloatDecoding(t, makeRandFloat(input, size), e)
-			fuzzDoubleDecoding(t, makeRandDouble(input, size), e)
+			fuzzFloatDecoding(t, fuzzing.MakeRandFloat(input, size), e)
+			fuzzDoubleDecoding(t, fuzzing.MakeRandDouble(input, size), e)
 		})
 	})
 }
@@ -121,9 +120,9 @@ func FuzzFloatEncoding(f *testing.F) {
 func FuzzIntEncoding(f *testing.F) {
 	f.Fuzz(func(t *testing.T, size int, input []byte) {
 		fuzzEncoding(func(e encoding.Encoding) {
-			fuzzInt32Decoding(t, makeRandInt32(input, size), e)
-			fuzzInt64Decoding(t, makeRandInt64(input, size), e)
-			fuzzInt96Decoding(t, makeRandInt96(input, size), e)
+			fuzzInt32Decoding(t, fuzzing.MakeRandInt32(input, size), e)
+			fuzzInt64Decoding(t, fuzzing.MakeRandInt64(input, size), e)
+			fuzzInt96Decoding(t, fuzzing.MakeRandInt96(input, size), e)
 		})
 	})
 }
@@ -304,109 +303,3 @@ func fuzzInt96Decoding(t *testing.T, input []deprecated.Int96, e encoding.Encodi
 		}
 	}
 }
-
-func makeRandBoolean(data []byte, count int) []bool {
-	if count < 1 {
-		return nil
-	}
-	src := rand.New(newByteSource(data))
-	b := make([]bool, count)
-	for i := 0; i < count; i++ {
-		b[i] = src.Int63()&0x01 == 1
-	}
-	return b
-}
-
-func makeRandFloat(data []byte, count int) []float32 {
-	if count < 1 {
-		return nil
-	}
-	src := rand.New(newByteSource(data))
-	f := make([]float32, count)
-	for i := 0; i < count; i++ {
-		f[i] = src.Float32()
-	}
-	return f
-}
-
-func makeRandDouble(data []byte, count int) []float64 {
-	if count < 1 {
-		return nil
-	}
-	src := rand.New(newByteSource(data))
-	f := make([]float64, count)
-	for i := 0; i < count; i++ {
-		f[i] = src.Float64()
-	}
-
-	return f
-}
-
-func makeRandInt32(data []byte, count int) []int32 {
-	if count < 1 {
-		return nil
-	}
-
-	src := rand.New(newByteSource(data))
-	a := make([]int32, count)
-	for i := 0; i < count; i++ {
-		a[i] = int32(src.Int63())
-	}
-	return a
-}
-
-func makeRandInt64(data []byte, count int) []int64 {
-	if count < 1 {
-		return nil
-	}
-
-	src := rand.New(newByteSource(data))
-	a := make([]int64, count)
-	for i := 0; i < count; i++ {
-		a[i] = src.Int63()
-	}
-	return a
-}
-
-func makeRandInt96(data []byte, count int) []deprecated.Int96 {
-	if count < 1 {
-		return nil
-	}
-
-	src := rand.New(newByteSource(data))
-	a := make([]deprecated.Int96, count)
-	for i := 0; i < count; i++ {
-		a[i] = deprecated.Int96{
-			uint32(src.Int63()),
-			uint32(src.Int63()),
-			uint32(src.Int63()),
-		}
-	}
-	return a
-}
-
-// byteSource is used to compose fuzz tests from a byte array.
-// This is to workaround the current stblib limitations.
-type byteSource struct {
-	*bytes.Reader
-}
-
-func newByteSource(data []byte) *byteSource {
-	return &byteSource{
-		Reader: bytes.NewReader(data),
-	}
-}
-
-func (s *byteSource) Uint64() uint64 {
-	var bytes [8]byte
-	if _, err := s.Read(bytes[:]); err != nil && !errors.Is(err, io.EOF) {
-		panic("byteSource: failed to read bytes")
-	}
-	return binary.BigEndian.Uint64(bytes[:])
-}
-
-func (s *byteSource) Int63() int64 {
-	return int64(s.Uint64() >> 1)
-}
-
-func (s *byteSource) Seed(seed int64) {}