--- a/internal/bits/min_arm64.s
+++ b/internal/bits/min_arm64.s
@@ -3,7 +3,7 @@
 #include "textflag.h"
 
 // func minBool(data []bool) bool
-TEXT ·minBool(SB), NOSPLIT, $0-25
+TEXT ·minBool(SB), NOSPLIT, $0-28
     MOVD data_base+0(FP), R0 // data base
     MOVD data_len+8(FP), R1 // length of data
     MOVD $ret+24(FP), R2 // address for result
@@ -23,7 +23,7 @@ loop:
     MOVB (R0), R4
     CMP $0, R4
     BEQ false
-    
+
     ADD $1, R0, R0
     SUB $1, R1, R1
 
@@ -59,22 +59,56 @@ TEXT ·minInt32(SB), NOSPLIT, $0-28
     MOVD data_len+8(FP), R1 // length of data
     MOVD $ret+24(FP), R2 // address for result
 
-    MOVD ZR, R3
-    MOVD $0x7FFFFFFF, R4
-    MOVD ZR, R6
-    MOVW$4, R7
+    MOVD R0, R6
+    ADD  R0, R1, R5
+
+    MOVD R1, R7
+
+    MOVD $0x7FFFFFFF, R8
+
+    B loopv
 
 loop:
-    MOVW (R6)(R0), R5
-    CMP R5, R4
-    CSEL LT, R4, R5, R4
+    CMP R0, R5
+    BEQ done
 
-    ADD $1, R3
-    MUL R7, R3, R6
+    MOVW.P 4(R0), R3
+    CMP R3, R8
+    CSEL LT, R8, R3, R8
 
-    CMP R3, R1
     BNE loop
 
 done:
-    MOVW R4, (R2)
+    MOVW R8, (R2)
     RET
+
+// a single int32 is stored in 32 bits (or 4 bytes)
+// armv8 vector sizes are 128bits, meaning we can store up to 4 int32 per vector.
+loopv:
+    CMP $8, R7
+    BLT loop
+
+    VLD1.P 32(R0), [V1.D2, V2.D2]
+    VUMIN V1.B16, V2.B16, V3.B16
+
+    VMOV V3.S[0], R3
+    VMOV V3.S[1], R4
+
+    CMP R3, R4
+    CSEL LT, R4, R3, R4
+
+    VMOV V3.S[2], R3
+    CMP R3, R4
+    CSEL LT, R4, R3, R4
+
+    VMOV V3.S[3], R3
+    CMP R3, R4
+    CSEL LT, R4, R3, R4
+
+    CMP R4, R8
+    CSEL LT, R8, R4, R8
+
+    SUB $8, R7
+    ADD $24, R5
+
+    B loopv