Opening this draft PR to show the work in progress to adapt the various SIMD optimization to ASIMD/NEON for ARM64 CPUs.

I will add the different benchmark as we go. But some promising results so far.

```
name                old time/op    new time/op      delta
MinBool/4KiB-10       1.28µs ± 0%      0.02µs ± 0%    -98.64%  (p=0.008 n=5+5)
MinBool/256KiB-10     82.1µs ± 0%       1.2µs ± 1%    -98.49%  (p=0.008 n=5+5)
MinBool/2048KiB-10     652µs ± 0%        11µs ± 6%    -98.38%  (p=0.016 n=4+5)

name                old speed      new speed        delta
MinBool/4KiB-10     3.20GB/s ± 0%  235.31GB/s ± 0%  +7261.29%  (p=0.008 n=5+5)
MinBool/256KiB-10   3.19GB/s ± 0%  212.08GB/s ± 1%  +6539.94%  (p=0.008 n=5+5)
MinBool/2048KiB-10  3.22GB/s ± 0%  198.28GB/s ± 6%  +6064.08%  (p=0.016 n=4+5)
```