@Pryz 2022-06-08
Closing for now. Will reopen once we are done with the refactoring of internal/bits.